refactor word substitution algorithms into independent files
